# webpage/platform-integration/docs/openapi-platform-mock.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Platform Integration Mock API",
    "description": "FastAPI 骨架，提供积分、授权、任务与审核回写的 Mock 接口供联调使用。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/authorizations/import": {
      "post": {
        "tags": [
          "authorizations"
        ],
        "summary": "批量导入授权",
        "operationId": "import_authorizations_api_v1_authorizations_import_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/authorizations/revoke": {
      "post": {
        "tags": [
          "authorizations"
        ],
        "summary": "吊销授权",
        "operationId": "revoke_api_v1_authorizations_revoke_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevokeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevokeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/authorizations/{template_id}": {
      "get": {
        "tags": [
          "authorizations"
        ],
        "summary": "查看模板授权名单",
        "operationId": "list_authorizations_api_v1_authorizations__template_id__get",
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Template Id"
            }
          },
          {
            "name": "tenant_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuthorizationRecord"
                  },
                  "title": "Response List Authorizations Api V1 Authorizations  Template Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/authorizations/{template_id}/logs": {
      "get": {
        "tags": [
          "authorizations"
        ],
        "summary": "查询授权使用日志",
        "operationId": "authorization_logs_api_v1_authorizations__template_id__logs_get",
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Template Id"
            }
          },
          {
            "name": "tenant_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Authorization Logs Api V1 Authorizations  Template Id  Logs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AuthorizationRecord": {
        "properties": {
          "authorization_id": {
            "type": "string",
            "title": "Authorization Id"
          },
          "template_id": {
            "type": "string",
            "title": "Template Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "授权状态",
            "default": "active"
          },
          "usage_limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Usage Limit",
            "description": "可用次数上限"
          },
          "used": {
            "type": "integer",
            "title": "Used",
            "description": "已使用次数",
            "default": 0
          },
          "valid_from": {
            "type": "string",
            "format": "date-time",
            "title": "Valid From"
          },
          "valid_to": {
            "type": "string",
            "format": "date-time",
            "title": "Valid To"
          }
        },
        "type": "object",
        "required": [
          "authorization_id",
          "template_id",
          "user_id",
          "tenant_id",
          "valid_from",
          "valid_to"
        ],
        "title": "AuthorizationRecord"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ImportRequest": {
        "properties": {
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "template_id": {
            "type": "string",
            "title": "Template Id"
          },
          "entries": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Entries",
            "description": "授权用户 ID 列表"
          },
          "usage_limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Usage Limit",
            "description": "次数限制"
          },
          "valid_days": {
            "type": "integer",
            "maximum": 365.0,
            "minimum": 1.0,
            "title": "Valid Days",
            "default": 30
          }
        },
        "type": "object",
        "required": [
          "tenant_id",
          "template_id",
          "entries"
        ],
        "title": "ImportRequest"
      },
      "ImportResponse": {
        "properties": {
          "template_id": {
            "type": "string",
            "title": "Template Id"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "created": {
            "type": "integer",
            "title": "Created"
          },
          "skipped": {
            "type": "integer",
            "title": "Skipped"
          },
          "batch_id": {
            "type": "string",
            "title": "Batch Id"
          }
        },
        "type": "object",
        "required": [
          "template_id",
          "total",
          "created",
          "skipped",
          "batch_id"
        ],
        "title": "ImportResponse"
      },
...
      "RevokeRequest": {
        "properties": {
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "template_id": {
            "type": "string",
            "title": "Template Id"
          },
          "user_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "User Ids"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "tenant_id",
          "template_id",
          "user_ids",
          "reason"
        ],
        "title": "RevokeRequest"
      },
      "RevokeResponse": {
        "properties": {
          "template_id": {
            "type": "string",
            "title": "Template Id"
          },
          "revoked": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Revoked"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "processed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Processed At"
          }
        },
        "type": "object",
        "required": [
          "template_id",
          "revoked",
          "reason",
          "processed_at"
        ],
        "title": "RevokeResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}